+++ Example/PaymentSheet Example/PaymentSheet Example/AppearancePlaygroundView.swift
@@ -120,6 +120,26 @@ struct AppearancePlaygroundView: View {
             set: { self.appearance.shadow.radius = $0 }
         )
 
+        let formInsetsTopBinding = Binding(
+            get: { self.appearance.formInsets.top },
+            set: { self.appearance.formInsets.top = $0 }
+        )
+
+        let formInsetsLeftBinding = Binding(
+            get: { self.appearance.formInsets.left },
+            set: { self.appearance.formInsets.left = $0 }
+        )
+
+        let formInsetsBottomBinding = Binding(
+            get: { self.appearance.formInsets.bottom },
+            set: { self.appearance.formInsets.bottom = $0 }
+        )
+
+        let formInsetsRightBinding = Binding(
+            get: { self.appearance.formInsets.right },
+            set: { self.appearance.formInsets.right = $0 }
+        )
+
         let sizeScaleFactorBinding = Binding(
             get: { self.appearance.font.sizeScaleFactor },
             set: { self.appearance.font.sizeScaleFactor = $0 }
@@ -318,6 +338,17 @@ struct AppearancePlaygroundView: View {
                             Text(String(format: "radius: %.1f", appearance.shadow.radius))
                             Slider(value: componentShadowRadiusBinding, in: 0...10, step: 0.5)
                         }
+                    }
+                    VStack {
+                        Text("formInsets")
+                        Stepper("top: \(Int(appearance.formInsets.top))",
+                                value: formInsetsTopBinding, in: 0...100)
+                        Stepper("left: \(Int(appearance.formInsets.left))",
+                                value: formInsetsLeftBinding, in: 0...100)
+                        Stepper("bottom: \(Int(appearance.formInsets.bottom))",
+                                value: formInsetsBottomBinding, in: 0...100)
+                        Stepper("right: \(Int(appearance.formInsets.right))",
+                                value: formInsetsRightBinding, in: 0...100)
                     }
                 }
                 Section(header: Text("Fonts")) {