+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/STPFixtures+PaymentSheet.swift
@@ -430,6 +430,7 @@ extension PaymentSheet.Appearance {
             offset: CGSize(width: 0, height: 2),
             radius: 4
         )
+        appearance.formInsets = UIEdgeInsets(top: 30, left: 50, bottom: 70, right: 10)
 
         // Customize the colors
         var colors = PaymentSheet.Appearance.Colors()